scrabble and player objects complete
